[Api][Shop] Add Checkout/CompletionTest
