+++ tests/Api/Responses/shop/checkout/completion/order_with_shippable_and_non_shippable_items.json
