--- tests/Api/Responses/shop/order_completion/payment_method_not_chosen_with_shipping_selected_state_response.json
@@ -1,13 +1,0 @@
-{
-    "@context": "\/api\/v2\/contexts\/ConstraintViolationList",
-    "@type": "ConstraintViolationList",
-    "hydra:title": "An error occurred",
-    "hydra:description": "Cannot complete as order is in a wrong state. Current: shipping_selected. Possible transitions: address, select_shipping, skip_payment, select_payment.",
-    "violations": [
-        {
-            "propertyPath": "",
-            "message": "Cannot complete as order is in a wrong state. Current: shipping_selected. Possible transitions: address, select_shipping, skip_payment, select_payment.",
-            "code": null
-        }
-    ]
-}
