--- tests/Api/Responses/shop/order_completion/shipping_method_not_chosen_response.json
@@ -1,13 +1,0 @@
-{
-    "@context": "\/api\/v2\/contexts\/ConstraintViolationList",
-    "@type": "ConstraintViolationList",
-    "hydra:title": "An error occurred",
-    "hydra:description": "Cannot complete as order is in a wrong state. Current: addressed. Possible transitions: address, skip_shipping, select_shipping.",
-    "violations": [
-        {
-            "propertyPath": "",
-            "message": "Cannot complete as order is in a wrong state. Current: addressed. Possible transitions: address, skip_shipping, select_shipping.",
-            "code": null
-        }
-    ]
-}
